refactor: 결제 모달 코드 분리
